combine `utils` into `task`.
also:
- add `tag` action and deprecate `tagging`
